client/requirements.txt
pillow==10.1.0
pytesseract==0.3.10
pyautogui==0.9.54
opencv-python==4.8.1.78
